requirements.txt
python-dotenv
telnetlib3
python-telegram-bot
python-dotenv
